got both api calls working and showing data. Need to work on UI, comments, refactoring, error handling, user feedback on error

+++ P_L_Tech_Test/Item.swift
@@ -8,7 +8,7 @@
 
 import Foundation
 struct Item: Decodable {
-	var id: Double
+	var id: Int
 	var title: String
 	var subtitle: String
 	var date: String
